chore: Rename EventsSearchButton to SearchToggleButton; Rename EventsSearchBar to ToggleSearcbBar
